@@ -1,14 +1,13 @@
-import { AnimatePresence, motion } from "framer-motion";
-import Link from "next/link";
-import Content from "../components/About/Content";
-import Footer from "../components/Footer";
-import ImageHover from "../components/ImageHover";
-import Navbar from "../components/Navbar/Navbar";
-import { fadeinAnimation } from "../constrants/fadeinAnimate";
+'use client';
+import { AnimatePresence, motion } from 'framer-motion';
+import Link from 'next/link';
+import Content from '../../../components/About/Content';
+import Footer from '../../../components/Footer';
+import ImageHover from '../../../components/ImageHover';
+import { fadeinAnimation } from '../../../constrants/fadeinAnimate';
 function About() {
   return (
     <div>
-      <Navbar />
       <div className="lg:px-8 pt-24 lg:pt-36 min-h-screen  md:px-5 relative ">
         <div className="max-w-7xl mx-auto mb-8">
           <div className="w-full h-full flex items-end px-1 ">
@@ -40,7 +39,7 @@ function About() {
                 <li>💎 Database: MySQL, MongoDB, Firebase,...</li>
               </ul>
               <p>
-                Also, i have some side experience like{" "}
+                Also, i have some side experience like{' '}
                 <span className="strong font-bold">Figma, Photoshop, Illustrator, After Effect</span>
               </p>
             </section>
@@ -48,7 +47,7 @@ function About() {
               <div className="sticky top-0">
                 <ImageHover
                   alt="avatar"
-                  src={"/avatar_1.jpg"}
+                  src={'/avatar_1.jpg'}
                   describe="This is me at FPT University 🎒"
                   height={2048}
                   width={1536}
@@ -65,7 +64,7 @@ function About() {
               <div className="grid grid-cols-3">
                 <ImageHover
                   alt="SolFest"
-                  src={"/Events/SolFest.jpg"}
+                  src={'/Events/SolFest.jpg'}
                   describe="SolFest 🔥🔥"
                   height={1577}
                   width={2738}
@@ -77,7 +76,8 @@ function About() {
                     initial="hidden"
                     whileInView="visible"
                     viewport={{ once: true }}
-                    className="text-xl bg-black/60 text-white w-fit px-4">
+                    className="text-xl bg-black/60 text-white w-fit px-4"
+                  >
                     <span className="opacity-30 text-white pl-4 ">&#47;&#47;</span> SolFest (07-11-2020 ➡️ 08-11-2020)
                   </motion.h5>
                   <motion.p
@@ -85,7 +85,8 @@ function About() {
                     initial="hidden"
                     whileInView="visible"
                     viewport={{ once: true }}
-                    className="px-3 md:px-0">
+                    className="px-3 md:px-0"
+                  >
                     ❇️ Solfest is a solution contest, where you can solve problems in different ways. I was a leader in
                     the
                     <span className="hightlight">TimeLỏd</span> and place the 2nd in the contest.
@@ -95,7 +96,8 @@ function About() {
                     initial="hidden"
                     whileInView="visible"
                     viewport={{ once: true }}
-                    className="px-3 md:px-0">
+                    className="px-3 md:px-0"
+                  >
                     ❇️ Through this competition, I have learned a lot about how to lead a team and how to work with
                     other. And, I was trained to create and present a good pitching.
                   </motion.p>
@@ -109,7 +111,8 @@ function About() {
                     initial="hidden"
                     whileInView="visible"
                     viewport={{ once: true }}
-                    className="text-xl bg-black/60 text-white w-fit px-4">
+                    className="text-xl bg-black/60 text-white w-fit px-4"
+                  >
                     <span className="opacity-30 text-white pl-4 ">&#47;&#47;</span> Techtips Copywriter (10-11-2020 ➡️
                     10-12-2020)
                   </motion.h5>
@@ -118,7 +121,8 @@ function About() {
                     initial="hidden"
                     whileInView="visible"
                     viewport={{ once: true }}
-                    className="px-3 md:px-0">
+                    className="px-3 md:px-0"
+                  >
                     ❇️ I&rsquo;m responsible for the IT section, through that, I have discovered a lot about IT section.
                     From AI to the new technologies,
                   </motion.p>
@@ -127,14 +131,15 @@ function About() {
                     initial="hidden"
                     whileInView="visible"
                     viewport={{ once: true }}
-                    className="px-3 md:px-0">
+                    className="px-3 md:px-0"
+                  >
                     ❇️ Like a charm, I have found some favourite technologies and frameworks that applied to many my
                     project till now .
                   </motion.p>
                 </section>
                 <ImageHover
                   alt="SolFest"
-                  src={"/Events/TechTips.png"}
+                  src={'/Events/TechTips.png'}
                   describe="TechTips 💪💪"
                   height={924}
                   width={1640}
@@ -147,7 +152,7 @@ function About() {
               <div className="grid grid-cols-3">
                 <ImageHover
                   alt="SolFest"
-                  src={"/Events/123host.png"}
+                  src={'/Events/123host.png'}
                   describe="123host 🌐"
                   height={1577}
                   width={2738}
@@ -159,7 +164,8 @@ function About() {
                     initial="hidden"
                     whileInView="visible"
                     viewport={{ once: true }}
-                    className="text-xl bg-black/60 text-white w-fit px-4">
+                    className="text-xl bg-black/60 text-white w-fit px-4"
+                  >
                     <span className="opacity-30 text-white pl-4 ">&#47;&#47;</span> Copywritter for 123host.vn
                     (20/11/2021 ➡️ 1/7/2022)
                   </motion.h5>
@@ -168,7 +174,8 @@ function About() {
                     initial="hidden"
                     whileInView="visible"
                     viewport={{ once: true }}
-                    className="px-3 md:px-0">
+                    className="px-3 md:px-0"
+                  >
                     ❇️ This is the first job that I have ever had. I&rsquo;m responsible for writing articles about
                     linux server&apos;services like webmail, SSH keys, nginx, etc.
                   </motion.p>
@@ -177,10 +184,15 @@ function About() {
                     initial="hidden"
                     whileInView="visible"
                     viewport={{ once: true }}
-                    className="px-3 md:px-0">
+                    className="px-3 md:px-0"
+                  >
                     ❇️ I have learned a lot about how to write a good article, how to use SEO to make the article, and
-                    how to work with unix system. You can view my section at{" "}
-                    <Link passHref className="hightlight" href={"https://123host.vn/community/user/ngocvlqt1995/tutorial/"}>
+                    how to work with unix system. You can view my section at{' '}
+                    <Link
+                      passHref
+                      className="hightlight"
+                      href={'https://123host.vn/community/user/ngocvlqt1995/tutorial/'}
+                    >
                       <span className="hightlight cursor-pointer hover:underline">123host.vn</span>
                     </Link>
                   </motion.p>
