@@ -1,5 +1,6 @@
-import Image from "next/image";
-import Link from "next/link";
+import Image from 'next/image';
+import Link from 'next/link';
+import { cn } from '../app/libs/tw';
 type Props = {
   src: string;
   alt: string;
@@ -14,12 +15,12 @@ function ImageHover(props: Props) {
   return (
     <Link href={props.href} passHref>
       <div
-        className={
-          "relative group mx-auto " +
-          (props.containerWidth ? props.containerWidth : "w-4/5") +
-          " " +
-          (props.containerHeight ? props.containerHeight : "h-full")
-        }>
+        className={cn([
+          'relative group mx-auto ',
+          props.containerWidth ? props.containerWidth : 'w-4/5',
+          props.containerHeight ? props.containerHeight : 'h-full',
+        ])}
+      >
         <div className="sticky top-20">
           <Image
             alt={props.alt}
@@ -28,7 +29,6 @@ function ImageHover(props: Props) {
             height={props.height}
             className="zoominAnimationOnHover object-contain"
             src={props.src}
-            priority={true}
           />
         </div>
         <h3 className="text-overlay opacity-0 group-hover:opacity-100 group-hover:-translate-y-5 bg-black/40 w-full">
