@@ -1,11 +1,7 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-  reactStrictMode: true,
   images: {
-    domains: ["images.unsplash.com"],
-  },
-  experimental: {
-    outputStandalone: true,
+    domains: ['images.unsplash.com'],
   },
 };
 
